Scrape Individual Player Statistics For One Team

Webscrape (using rvest) all individual player statistics of a given team for the current season

Usage

scrapeIndivStatsByTeam(
  team,
  games = "all",
  players = "goalies",
  gender = "women"
)

Arguments

team

ECAC team to scrape data for

games

collecting data for 'all' (default), 'conference', or 'nonconference' games. Currently no support for 'nonconference' games

players

type of players stats to collect, 'goalies' (default) or 'skaters'

gender

'women' (default) or 'men'

Value

data frame of individual player statistics for one ECAC team

Examples

scrapeIndivStatsByTeam(team="brown")
#> # A tibble: 3 × 15
#>     No. Name        Year  Position GamesPlayed GamesStarted Minutes GoalsAgainst
#>   <int> <chr>       <chr> <chr>          <int>        <int> <chr>          <int>
#> 1     1 Grace  Ked… Fr    G                  5            0 27:49              1
#> 2    31 Kaley  Doy… So    G                 23           16 1246:05           44
#> 3    53 Hayley  Kl… Sr    G                 10            8 469:22            39
#> # … with 7 more variables: GoalsAgainstAverage <dbl>, Saves <int>,
#> #   SavePercentage <dbl>, Wins <int>, Losses <int>, Ties <int>,
#> #   WinPercentage <dbl>
scrapeIndivStatsByTeam(team="clarkson", games="conference", players="skaters", gender="men")
#> # A tibble: 26 × 17
#>      No. Name     Year  Position GamesPlayed Goals Assists Points PenaltyMinutes
#>    <int> <chr>    <chr> <chr>          <int> <int>   <int>  <int>          <int>
#>  1     2 Lukas  … Sr    D                 18     2       9     11             10
#>  2     3 Kaelan … So    D                 20     0       3      3              4
#>  3     4 Brian  … Sr    D                 13     2       2      4              2
#>  4     5 Michael… Sr    D                 21     0       3      3             10
#>  5     6 Jack  J… Sr    F                 23     2      19     21             18
#>  6     8 Chris  … Sr    F                 19     1       0      1             10
#>  7     9 Jamie  … Sr    F                 12     2       0      2              4
#>  8    10 Nick  W… Jr    F                 NA    NA      NA      0             NA
#>  9    11 Alex  C… So    F                 23    14      15     29              2
#> 10    12 Caden  … Fr    D                 NA    NA      NA      0             NA
#> # … with 16 more rows, and 8 more variables: PlusMinus <int>,
#> #   PowerPlayGoals <int>, ShortHandedGoals <int>, EmptyNetGoals <int>,
#> #   GameWinningGoals <int>, GameTyingGoals <int>, HatTricks <int>, Shots <int>